split tests in different modules
